Change filename for database to match regex
Before, anytime a change was made to the spreadsheet (and by extension
the filename), the python would stop working. This happened because it
looked for a specific string. Now, it matches it to this regex:
"scouting.*/.xlsm", basically looking for scouting.xlsm with anything,
like the date and time, in between 'scouting' and '.'.